helpers: move ELF and IDA helpers out of script and into common module
